requirements.txt
torch==1.10.1
torchvision==0.11.2
cvbase
imageio
matplotlib
numpy
opencv-python
Pillow
PyYAML
scikit-image
scipy
tensorboardX
imageio-ffmpeg
Cython
colorama
requests
fire
numba
h5py
tqdm